% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_cnn_model}
\alias{build_cnn_model}
\title{Build CNN Model}
\usage{
build_cnn_model(input_shape, output_units)
}
\arguments{
\item{input_shape}{Integer vector. Shape of the input data.}

\item{output_units}{Integer. Number of output units.}
}
\value{
A compiled CNN model.
}
\description{
Constructs a CNN model using \code{keras}.
}
