% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{calibrate_sir}
\alias{calibrate_sir}
\title{Calibrate SIR Parameters (one-step)}
\usage{
calibrate_sir(
  daily_cases,
  population_size,
  recovery_rate,
  model_dir = NULL,
  auto_init = TRUE
)
}
\description{
Calibrate SIR Parameters (one-step)
}
