% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{check_model_status}
\alias{check_model_status}
\title{Check model status}
\usage{
check_model_status()
}
\value{
A list describing model load status and Python config.
}
\description{
Check model status
}
