% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{cleanup_bilstm_model}
\alias{cleanup_bilstm_model}
\title{Free Python-side model and scalers}
\usage{
cleanup_bilstm_model()
}
\value{
NULL (invisible)
}
\description{
Free Python-side model and scalers
}
