% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir.R
\name{.check_conda_env_health}
\alias{.check_conda_env_health}
\title{Check package conda environment health}
\usage{
.check_conda_env_health(envname)
}
\description{
Check package conda environment health
}
\keyword{internal}
