% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{.ensure_python_ready}
\alias{.ensure_python_ready}
\title{Ensure Python is usable and required modules are present.
\itemize{
\item Respects RETICULATE_PYTHON if the user has set it.
\item Otherwise creates/uses a private virtualenv 'epiworldRcalibrate'
\item Installs numpy, joblib, torch (CPU) if missing.
}}
\usage{
.ensure_python_ready()
}
\description{
Ensure Python is usable and required modules are present.
\itemize{
\item Respects RETICULATE_PYTHON if the user has set it.
\item Otherwise creates/uses a private virtualenv 'epiworldRcalibrate'
\item Installs numpy, joblib, torch (CPU) if missing.
}
}
\keyword{internal}
