% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir.R
\name{.required_python_modules}
\alias{.required_python_modules}
\title{Return required Python module names}
\usage{
.required_python_modules()
}
\description{
Return required Python module names
}
\keyword{internal}
