% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir.R
\name{.required_python_packages}
\alias{.required_python_packages}
\title{Return required Python package names for installation}
\usage{
.required_python_packages()
}
\description{
Return required Python package names for installation
}
\keyword{internal}
