% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir.R
\name{.verify_python_imports}
\alias{.verify_python_imports}
\title{Verify module imports using reticulate's active Python}
\usage{
.verify_python_imports(modules = .required_python_modules())
}
\description{
Verify module imports using reticulate's active Python
}
\keyword{internal}
