% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{estimate_sir_parameters}
\alias{estimate_sir_parameters}
\title{Estimate SIR Parameters from 61-day incidence}
\usage{
estimate_sir_parameters(daily_cases, population_size, recovery_rate)
}
\value{
named numeric vector: ptran, crate, R0
}
\description{
Estimate SIR Parameters from 61-day incidence
}
