% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate the CNN Model}
\usage{
evaluate_model(model, test_data, theta)
}
\arguments{
\item{model}{Keras model. The trained model to be evaluated.}

\item{test_data}{List. A list containing test data (\code{x} and \code{y}).}

\item{theta}{\code{data.table}. The parameters corresponding to the simulations.}
}
\value{
A list containing:
\describe{
\item{pred}{The predicted values.}
\item{MAEs}{The Mean Absolute Errors for each output variable.}
}
}
\description{
Evaluates the trained model on test data, makes predictions, and calculates Mean Absolute Error (MAE).
}
