% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_null_infected.R
\name{filter_non_null_infected}
\alias{filter_non_null_infected}
\title{Filter Non-Null and Valid Matrices}
\usage{
filter_non_null_infected(matrices)
}
\arguments{
\item{matrices}{A list of matrices to be filtered.}
}
\value{
A list with two elements:
\item{matrices}{The filtered list of valid matrices.}
\item{N}{The number of valid matrices remaining after filtering.}
}
\description{
This function filters a list of matrices, excluding those that contain errors or missing values.
}
\details{
The function performs the following checks on each matrix in the input list:
\enumerate{
\item Excludes matrices that inherit an "error" class.
\item Excludes matrices that contain any \code{NA} values.
}

The function then returns the filtered matrices and their count.
}
