% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_theta.R
\name{generate_theta}
\alias{generate_theta}
\title{Generate Theta Parameters}
\usage{
generate_theta(N, n)
}
\arguments{
\item{N}{Integer. Number of datasets to generate.}

\item{n}{Integer. Size of the population.}
}
\value{
A \code{data.table} containing theta parameters.
}
\description{
Generates the theta parameters for the simulation pipeline.
}
