% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{get_bilstm_model_dir}
\alias{get_bilstm_model_dir}
\title{Get current model directory path}
\usage{
get_bilstm_model_dir()
}
\value{
Character string with current model directory, or NULL if not loaded
}
\description{
Get current model directory path
}
