% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{is_bilstm_loaded}
\alias{is_bilstm_loaded}
\title{Check if BiLSTM Model is Loaded}
\usage{
is_bilstm_loaded()
}
\value{
Logical indicating if the model is currently loaded
}
\description{
Check if BiLSTM Model is Loaded
}
