% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pipeline.R
\name{main_pipeline}
\alias{main_pipeline}
\title{Main Pipeline}
\usage{
main_pipeline(N, n, ndays, ncores)
}
\arguments{
\item{N}{Integer. Number of datasets.}

\item{n}{Integer. Size of the population.}

\item{ndays}{Integer. Number of simulation days.}

\item{ncores}{Integer. Number of cores for parallelization.}
}
\description{
Executes the full pipeline: simulation, data preparation, and modeling.
}
