% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot Results of Model Predictions}
\usage{
plot_results(pred, test_data, theta, MAEs, N, N_train)
}
\arguments{
\item{pred}{\code{data.table}. The predicted values from the model.}

\item{test_data}{List. A list containing test data (\code{x} and \code{y}).}

\item{theta}{\code{data.table}. The parameters corresponding to the simulations.}

\item{MAEs}{Numeric vector. The Mean Absolute Errors for each output variable.}

\item{N}{Integer. Total number of simulations.}

\item{N_train}{Integer. Number of training examples.}
}
\value{
Generates plots (displayed on the active graphics device).
}
\description{
Generates plots comparing predicted and observed values, including boxplots and scatter plots with MAE labels.
}
