% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{predict_sir_bilstm}
\alias{predict_sir_bilstm}
\title{Fast SIR parameter prediction (back-compat)}
\usage{
predict_sir_bilstm(time_series, n, recov)
}
\arguments{
\item{time_series}{Numeric vector of length 61 (incidence)}

\item{n}{Numeric (>0). Population size}

\item{recov}{Numeric (>0). Recovery rate}
}
\value{
Named numeric vector \code{c(ptran, crate, R0)}
}
\description{
Fast SIR parameter prediction (back-compat)
}
