% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for model analysis}
\usage{
prepare_data(m, max_days = 50)
}
\arguments{
\item{m}{An epiworldR model object.}

\item{max_days}{The maximum number of days to include in the output.}
}
\value{
An array containing processed simulation data.
}
\description{
This function processes simulation output from epiworldR and prepares it for use in
predictive modeling or analysis. It handles incidence data, reproductive numbers,
and generation times.
}
