% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_for_tensorflow}
\alias{prepare_data_for_tensorflow}
\title{Prepare Data for TensorFlow}
\usage{
prepare_data_for_tensorflow(matrices, N)
}
\arguments{
\item{matrices}{List of simulation results.}

\item{N}{Integer. Number of datasets.}
}
\value{
A 3D array for TensorFlow input.
}
\description{
Converts simulation results into TensorFlow-compatible arrays.
}
