% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_infected.R
\name{prepare_data_infected}
\alias{prepare_data_infected}
\title{Prepare Data for TensorFlow Model: Infected Cases}
\usage{
prepare_data_infected(data)
}
\arguments{
\item{data}{A numeric vector representing the number of infected cases over time (e.g., daily incidence).}
}
\value{
A reshaped array of processed data suitable for TensorFlow models. If an error occurs, it returns the error object.
}
\description{
Prepare Data for TensorFlow Model: Infected Cases
}
