% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{prepare_data_infections_only.default}
\alias{prepare_data_infections_only.default}
\title{Default method for infections-only data}
\usage{
\method{prepare_data_infections_only}{default}(m, max_days = 50, ...)
}
\arguments{
\item{m}{A data.table containing incidence data.}

\item{max_days}{The maximum number of days to include in the output.}

\item{...}{Additional parameters.}
}
\value{
An array containing processed infection data.
}
\description{
Handles generic data for infections-only processing.
}
