% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{prepare_data_infections_only.epiworld_model}
\alias{prepare_data_infections_only.epiworld_model}
\title{Method for epiworld model}
\usage{
\method{prepare_data_infections_only}{epiworld_model}(m, ...)
}
\arguments{
\item{m}{An epiworldR model object.}

\item{...}{Additional parameters.}
}
\description{
Prepares infections-only data specifically for an epiworldR model.
}
