% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{prepare_data_infections_only}
\alias{prepare_data_infections_only}
\title{Prepare infections-only data}
\usage{
prepare_data_infections_only(m, ...)
}
\arguments{
\item{m}{An epiworldR model object or data.table.}

\item{...}{Additional parameters passed to methods.}
}
\value{
An array containing processed infection data.
}
\description{
Processes incidence data from epiworldR to extract only infection counts.
}
