% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_simulation.R
\name{read_training_data}
\alias{read_training_data}
\title{Read externally generated training data}
\usage{
read_training_data(parameter_file, observation_file, type = "generic")
}
\arguments{
\item{parameter_file}{CSV file containing one parameter row per simulation.}

\item{observation_file}{CSV file containing one observed time series per row.}

\item{type}{Optional epidemiological model label.}
}
\description{
Read externally generated training data
}
