% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Run SIR Model Simulations}
\usage{
run_simulations(N, n, ndays, ncores, theta, seeds, path)
}
\arguments{
\item{N}{Integer. The number of simulations to run.}

\item{n}{Integer. The population size for each simulation.}

\item{ndays}{Integer. The number of days to simulate.}

\item{ncores}{Integer. The number of cores to use for parallel processing.}

\item{theta}{\code{data.table}. The parameters for the simulations.}

\item{seeds}{Integer vector. Random seeds for each simulation.}
}
\value{
A list containing the simulation results as matrices.
}
\description{
Runs SIR model simulations using the \code{epiworldR} package in parallel.
}
