% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_simulation.R
\name{seir_simulator}
\alias{seir_simulator}
\title{Create the current epiworldR SEIR simulator}
\usage{
seir_simulator(ndays = 365, priors = NULL)
}
\arguments{
\item{ndays}{Number of returned incidence days.}

\item{priors}{Optional named list overriding the default prior ranges.}
}
\description{
Create the current epiworldR SEIR simulator
}
