% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{show_model_search_paths}
\alias{show_model_search_paths}
\title{Show searched model directory locations}
\usage{
show_model_search_paths()
}
\value{
Character vector of searched paths
}
\description{
Utility function to help debug model loading issues by showing
all the directories that would be searched for model files.
}
