% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_simulation.R
\name{simulate_from_parameters}
\alias{simulate_from_parameters}
\title{Simulate one curve from a simulator}
\usage{
simulate_from_parameters(simulator, parameters, seed = 1)
}
\arguments{
\item{simulator}{A `simulator` object.}

\item{parameters}{Named list or one-row data frame of parameter values.}

\item{seed}{Simulation seed.}
}
\description{
Simulate one curve from a simulator
}
