% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split Data into Training and Test Sets}
\usage{
split_data(arrays, theta, N)
}
\arguments{
\item{arrays}{Array. The 3D array prepared for TensorFlow.}

\item{theta}{\code{data.table}. The parameters corresponding to the simulations.}

\item{N}{Integer. The number of simulations.}
}
\value{
A list containing:
\describe{
\item{train}{A list with training data (\code{x} and \code{y}).}
\item{test}{A list with test data (\code{x} and \code{y}).}
}
}
\description{
Splits the data into training and test sets for model training and evaluation.
}
