% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train the CNN Model}
\usage{
train_model(model, train_data, epochs = epochs, verbose = verbose)
}
\arguments{
\item{model}{Keras model. The compiled model to be trained.}

\item{train_data}{List. A list containing training data (\code{x} and \code{y}).}

\item{epochs}{Integer. The number of epochs for training.}

\item{verbose}{Integer. 0 shows the result of training and 2 doesnt.}
}
\value{
The trained model (updated in place).
}
\description{
Trains the CNN model on the provided training data.
}
