% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cnn_model.R
\name{build_cnn_model}
\alias{build_cnn_model}
\title{Build a Convolutional Neural Network Model}
\usage{
build_cnn_model(input_shape, output_units)
}
\arguments{
\item{input_shape}{Integer vector. The shape of the input data (excluding batch size).}

\item{output_units}{Integer. The number of output units (number of output variables).}
}
\value{
A compiled Keras model ready for training.
}
\description{
Constructs and compiles a CNN model using the \code{keras3} package.
}
