% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir.R
\name{calibrate_sir}
\alias{calibrate_sir}
\title{Calibrate SIR Model Predictions}
\usage{
calibrate_sir(data)
}
\arguments{
\item{data}{A numeric matrix or array containing the input test data.}
}
\value{
A list containing a \code{data.table} of predicted values.
}
\description{
Generates predictions for input test data using a pre-trained CNN model.
}
