% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir.R
\name{calibrate_sir}
\alias{calibrate_sir}
\title{Predict Parameters Using a CNN Model
Generates predictions for input test data using a pre-trained Convolutional Neural Network (CNN) model.}
\usage{
calibrate_sir(data)
}
\arguments{
\item{data}{A numeric matrix or array containing the input test data. The function determines which model to load based on the number of columns in \code{data} (expects either 30 or 60).}
}
\value{
A list containing:
\describe{
\item{pred}{A \code{data.table} of predicted values with the following columns:
\describe{
\item{\code{preval}}{Predicted prevalence.}
\item{\code{crate}}{Predicted case rate.}
\item{\code{ptran}}{Predicted transmission probability.}
\item{\code{prec}}{Predicted precision.}
}
}
}
}
\description{
This function loads a specific Keras CNN model based on the length of the input data and uses it to make predictions. The predicted values are returned as a \code{data.table} with standardized column names.
}
\details{
The function determines which pre-trained CNN model to load based on the number of features (columns) in the input \code{data}. If \code{data} has 30 columns, it loads the \code{sir30-cnn.keras} model; if it has 60 columns, it loads the \code{sir60-cnn.keras} model. Ensure that the input data matches one of these expected formats to avoid errors.
}
