% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir.R
\docType{package}
\name{epiworldRcalibrate-package}
\alias{epiworldRcalibrate}
\alias{epiworldRcalibrate-package}
\title{epiworldRcalibrate: Fast and Effortless Calibration of Agent-Based Models using Machine Learning}
\description{
Provides tools and pre-trained Machine Learning [ML] models for calibration of Agent-Based Models [ABMs] built with the R package 'epiworldR'. Implements methods described in Najafzadehkhoei, Vega Yon, Modenesi, and Meyer (2025) \doi{10.48550/arXiv.2509.07013}. Users can automatically calibrate ABMs in seconds with pre-trained ML models, effectively focusing on simulation rather than calibration. Bridges a gap by allowing public health practitioners to run their own ABMs without the advanced technical expertise often required by calibration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sima-njf.github.io/epiworldRcalibrate/}
  \item \url{https://github.com/sima-njf/epiworldRcalibrate}
  \item Report bugs at \url{https://github.com/sima-njf/epiworldRcalibrate/issues}
}

}
\author{
\strong{Maintainer}: Sima Najafzadehkhoei \email{sima.njf@utah.edu} (\href{https://orcid.org/0009-0002-6253-2910}{ORCID})

Authors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})
  \item Bernardo Modenesi \email{bmodenesi@gmail.com}
}

Other contributors:
\itemize{
  \item Centers for Disease Control and Prevention (Award number 1U01CK000585; 75D30121F00003) [funder]
}

}
\keyword{internal}
