% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\docType{package}
\name{epiworldRcalibrate-package}
\alias{epiworldRcalibrate}
\alias{epiworldRcalibrate-package}
\title{epiworldRcalibrate: Calibration with LSTM models}
\description{
Provides functions for calibration using Long-short term memory (LSTM). The package includes the pre-treined model which can be used directly on your data to do the calibration.
}
\author{
\strong{Maintainer}: Sima Najafzadehkhoei \email{sima.njf@utah.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
