% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_null.R
\name{filter_non_null}
\alias{filter_non_null}
\title{Filter Non-Null Simulation Results}
\usage{
filter_non_null(matrices, theta)
}
\arguments{
\item{matrices}{List. The list of simulation result matrices.}

\item{theta}{\code{data.table}. The parameters corresponding to the simulations.}
}
\value{
A list containing:
\describe{
\item{matrices}{The filtered list of matrices.}
\item{theta}{The filtered parameters.}
\item{N}{The updated number of simulations.}
}
}
\description{
Filters out simulations that resulted in errors or contain \code{NA} values.
}
