% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_theta.R
\name{generate_theta}
\alias{generate_theta}
\title{Generate Theta Parameters for SIR Model Simulation}
\usage{
generate_theta(N, n)
}
\arguments{
\item{N}{Integer. The number of parameter sets to generate.}

\item{n}{Integer. The population size for each simulation.}
}
\value{
A data.table containing the generated parameters.
}
\description{
Generates a data table of parameters needed to simulate SIR models.
}
