% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for TensorFlow Model: General SIR Data Preparation}
\usage{
prepare_data(m, max_days = max_days)
}
\arguments{
\item{m}{An \code{epiworldR} model object.}

\item{max_days}{The maximum number of days to consider for data preparation. Defaults to 50.}
}
\value{
A reshaped array of processed data suitable for TensorFlow models.
If an error occurs, it returns the error object.
}
\description{
Prepare Data for TensorFlow Model: General SIR Data Preparation
}
