% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_tensorflow.R
\name{prepare_data_for_tensorflow}
\alias{prepare_data_for_tensorflow}
\title{Prepare Data for TensorFlow}
\usage{
prepare_data_for_tensorflow(matrices, N)
}
\arguments{
\item{matrices}{List. The list of simulation result matrices.}

\item{N}{Integer. The number of simulations.}
}
\value{
A 3D array suitable for input into TensorFlow models.
}
\description{
Converts a list of matrices into a 3D array suitable for TensorFlow input.
}
