% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_data.R
\name{preprocessing_data}
\alias{preprocessing_data}
\title{Preprocesses Data for Model Input}
\usage{
preprocessing_data(data)
}
\arguments{
\item{m}{A data structure containing the input data. Assumes a named list with an element \code{Infected}.}
}
\value{
A reshaped array containing preprocessed data. If an error occurs, the function returns an error object.
}
\description{
This function preprocesses data for use in modeling. It handles missing values, filters data
based on the maximum number of days, and reshapes the output into a structured array format.
}
\details{
The function performs the following steps:
\itemize{
\item Converts the input data to a \code{data.table} format.
\item Filters rows based on \code{max_days}.
\item Fills missing values using the last observation carried forward ("locf").
\item Calculates the difference between consecutive rows after the first observation.
\item Reshapes the processed data into a 3D array.
}

If any error occurs during processing, the function catches the error and returns it.
}
