% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Run Simulations}
\usage{
run_simulations(N, n, ndays, ncores, theta, seeds)
}
\arguments{
\item{N}{Integer. Number of simulations to run.}

\item{n}{Integer. Size of the population.}

\item{ndays}{Integer. Number of days to simulate.}

\item{ncores}{Integer. Number of cores for parallel execution.}

\item{theta}{A \code{data.table} of theta parameters.}

\item{seeds}{Integer vector of seeds for reproducibility.}
}
\value{
A list of simulation results.
}
\description{
Runs the SIR-CONN simulations based on theta parameters.
}
