% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sir_bilstm.R
\name{show_preprocessing}
\alias{show_preprocessing}
\title{Demonstrate preprocessing (not required for prediction)}
\usage{
show_preprocessing(raw_counts)
}
\description{
Demonstrate preprocessing (not required for prediction)
}
